+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.19.1 (2025-04-02)
+* Upgraded Collections dependency
+* Support named arguments for attributes
 * Use null tagName for Each and Elements Components
 * Type hint $el instances in Components